refactor: Migrate AppointmentRescheduleScreen to MVI
This commit refactors the `AppointmentRescheduleScreen` by:
- Deleting the old `AppointmentRescheduleScreen.kt` file.
- Renaming `AppointmentRescheduleContent.kt` to `AppointmentRescheduleScreen.kt` and moving it to the `reschedule` subpackage.
- Updating the new `AppointmentRescheduleScreen` to adopt the Model-View-Intent (MVI) architecture.

**Key changes in the new `AppointmentRescheduleScreen`:**
- Integrates `RescheduleViewModel` (using Koin for dependency injection).
- Manages UI state (`RescheduleState`) and handles UI events (`RescheduleEvent`).
- Uses `LaunchedUiEffectHandler` to manage side effects (`RescheduleEffect`) like navigation and showing messages.
- Displays success and error messages using `SuccessesDialog` and `DialogWithIcon` respectively.
- Makes `AppointmentRescheduleContent` composable private as it's only used within this screen.
- Updates preview function `AppointmentRescheduleContentPreview` to be private.

**Affected files:**
- **features/appointments/src/patient/kotlin/eg/edu/cu/csds/icare/appointment/screen/AppointmentRescheduleScreen.kt:** (deleted)
- **features/appointments/src/patient/kotlin/eg/edu/cu/csds/icare/appointment/screen/reschedule/AppointmentRescheduleScreen.kt:** (renamed from `features/appointments/src/patient/kotlin/eg/edu/cu/csds/icare/appointment/screen/AppointmentRescheduleContent.kt`)
    - Updated package declaration.
    - Implemented MVI pattern with `RescheduleViewModel`.
    - Added `LaunchedUiEffectHandler` for side effects.
    - Changed visibility of internal composables and preview function.
